+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/auth/core/CustomOAuth2UserServiceImpl.java
@@ -15,11 +15,10 @@ import jakarta.servlet.http.HttpSession;
 import kernel.jdon.member.domain.Member;
 import kernel.jdon.member.domain.MemberRole;
 import kernel.jdon.member.domain.SocialProviderType;
-import kernel.jdon.moduleapi.domain.auth.dto.JdonOAuth2User;
-import kernel.jdon.moduleapi.domain.auth.dto.SessionUserInfo;
 import kernel.jdon.moduleapi.domain.auth.error.AuthErrorCode;
 import kernel.jdon.moduleapi.domain.member.core.MemberReader;
 import kernel.jdon.moduleapi.domain.member.core.MemberStore;
+import kernel.jdon.moduleapi.global.dto.SessionUserInfo;
 import kernel.jdon.moduleapi.global.exception.AuthException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
