+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/auth/core/JdonOAuth2User.java
@@ -1,4 +1,4 @@
-package kernel.jdon.moduleapi.domain.auth.dto;
+package kernel.jdon.moduleapi.domain.auth.core;
 
 import java.util.Collection;
 import java.util.Map;
