+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/auth/core/OAuth2Strategy.java
@@ -3,8 +3,8 @@ package kernel.jdon.moduleapi.domain.auth.core;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 
 import kernel.jdon.member.domain.SocialProviderType;
-import kernel.jdon.moduleapi.domain.auth.dto.SessionUserInfo;
 import kernel.jdon.moduleapi.domain.auth.error.AuthErrorCode;
+import kernel.jdon.moduleapi.global.dto.SessionUserInfo;
 import kernel.jdon.moduleapi.global.exception.AuthException;
 
 public interface OAuth2Strategy {
