+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/member/application/MemberFacade.java
@@ -3,11 +3,11 @@ package kernel.jdon.moduleapi.domain.member.application;
 import org.springframework.stereotype.Service;
 
 import kernel.jdon.moduleapi.domain.auth.core.CustomOAuth2UserServiceImpl;
-import kernel.jdon.moduleapi.domain.auth.dto.SessionUserInfo;
 import kernel.jdon.moduleapi.domain.auth.error.AuthErrorCode;
 import kernel.jdon.moduleapi.domain.member.core.MemberCommand;
 import kernel.jdon.moduleapi.domain.member.core.MemberInfo;
 import kernel.jdon.moduleapi.domain.member.core.MemberService;
+import kernel.jdon.moduleapi.global.dto.SessionUserInfo;
 import kernel.jdon.moduleapi.global.exception.ApiException;
 import lombok.RequiredArgsConstructor;
