+++ module-api/src/main/java/kernel/jdon/moduleapi/global/config/auth/JdonOAuth2AuthenticationSuccessHandler.java
@@ -13,8 +13,8 @@ import org.springframework.stereotype.Component;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import kernel.jdon.moduleapi.domain.auth.dto.JdonOAuth2User;
 import kernel.jdon.member.domain.MemberRole;
+import kernel.jdon.moduleapi.domain.auth.core.JdonOAuth2User;
 import kernel.jdon.moduleapi.domain.auth.util.CryptoManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
